# Fix ReduceMean/ReduceSum performance by specializing Eigen implementations for common shapes