--- a/caffe2/operators/reduce_ops.h
+++ b/caffe2/operators/reduce_ops.h
@@ -74,6 +74,7 @@ class ReduceOpBase : public Operator<Context> {
         Y->template mutable_data<T>(),
         Y_size,
         axes_,
+        y_dims,
         keepdims_);
   }
 
@@ -85,6 +86,7 @@ class ReduceOpBase : public Operator<Context> {
       T* Y_data,
       const TIndex Y_size,
       vector<int>& axes,
+      vector<TIndex>& Y_dims,
       int keepdims) = 0;
 
  private:
@@ -108,6 +110,7 @@ class ReduceSumOp : public ReduceOpBase<T, Context> {
       T* Y_data,
       const TIndex Y_size,
       vector<int>& axes,
+      vector<TIndex>& Y_dims,
       int keepdims) override;
 };
 
@@ -127,6 +130,7 @@ class ReduceMeanOp : public ReduceOpBase<T, Context> {
       T* Y_data,
       const TIndex Y_size,
       vector<int>& axes,
+      vector<TIndex>& Y_dims,
       int keepdims) override;
 };