…ommon shapes


# Perf analysis

Setup: Intel(R) Xeon(R) CPU E5-2680 v3 @ 2.50GHz

Reduce [2, 100100] tensor to [100100]. This is a shape seen in the LATTE output projection. Mean should amount to 200 kFLOPs (add + divide per column) and sum should be 100 kFLOPs (add per column)

**Before**

`benchReduceMean`

- 12.4465 ms/iter
- 16.7 MFLOP/s throughput

`benchReduceSum`

- 12.3889 ms/iter
- 8 MFLOP/s thoroughput

**After**

`benchReduceMean`

- 0.218622 ms/iter
- 914 MFLOP/s

`benchReduceSum`

- 0.209059 ms/iter
- 478 MFLOP/s